Move deepGet to shared helper
